Fix for incorrect castle move parsing
The castles moves were incorrectly parsed, (short for long, long for
short castle). This caused the rooks to end up in the wrong place,
causing exceptions later.

+++ Chess/Move/Move.hs
@@ -64,8 +64,8 @@ parserMove b = do
               else Nothing
       cstl  = if Just King == pieceAt b f
               then case hDiff f t of
-                2  -> return Long
-                (-2) -> return Short
+                2  -> return Short
+                (-2) -> return Long
                 _  -> Nothing
               else Nothing
   return
